feat: Add is_email_verified field to UserAdmin list_display
This commit adds the "is_email_verified" field to the "list_display" attribute in the UserAdmin class of the admin.py file. By including this field, the email verification status of users will be displayed in the admin interface.

This change enhances the functionality of the application by providing visibility into the email verification status of users in the admin interface.