refactor: Add payment method field to Subscription model
This commit adds a new field, `payment_method`, to the `Subscription` model in the `payment` app. The field is a `CharField` with a maximum length of 50 and has two choices: "paypal" and "credit_card". This change allows users to specify their preferred payment method when creating a subscription.

These changes improve the functionality and flexibility of the subscription feature by providing users with the option to choose their payment method.