refactor: Update admin site headers and titles
The commit updates the `ISPECO_Core/urls.py` file to modify the headers and titles of the admin site. This change sets the site header to "ISPECO Administration", the site title to "ISPECO Site Admin", and the index title to "Welcome to ISPECO Admin Panel". This improves the branding and user experience of the admin panel.